// src/policy/domain/enum/documentActorRelationship.enum.ts
enum DocumentActorRelationship {
	AUTHOR = "author",
	NAMED_INDIVIDUAL = "named_individual",
	TARGET_HANDLER = "target_handler",
	AUTHORIZED_CUSTODIAN = "authorized_custodian",
	UNIT_HEAD = "unit_head",
	DELEGATED_UNIT_HEAD = "delegated_unit_head",
	GUEST_READER = "guest_reader",
	PRIMARY_AUTHORIZING_DESK = "primary_authorizing_desk",
}

export { DocumentActorRelationship };