// src/policy/domain/enum/documentGovernanceAction.enum.ts
enum DocumentGovernanceAction {
	ASSIGN_SENSITIVITY = "assign_sensitivity",
	CHANGE_SENSITIVITY = "change_sensitivity",
	DISCOVER = "discover",
	VIEW = "view",
	FORWARD = "forward",
	EXPORT = "export",
	PRINT = "print",
	ATTACH = "attach",
	MANAGE_CC = "manage_cc",
	RENDER_CC_HEADER = "render_cc_header",
}

export { DocumentGovernanceAction };